Vulkan: keep UV clamping; promote vTexLimits and vUV to highp
Restore clamp_coord() on the nearest-neighbour path and fix the actual
cause of the Android texture corruption (#992): vTexLimits is declared
mediump, and the rhi disables clamping under a texture window by sending
UINT16_MAX limits. Mobile drivers that honour RelaxedPrecision with
16-bit integers fold 65535 to -1, so clamp(uv, 0, -1) pins every texel
to base - 1. Desktop drivers keep the value at 32 bits, which is why the
clamp is a no-op there and the clamp itself was never the problem.
Declaring the limits highp (and UVRange in the vertex stage) fixes
nearest, every filtered path and the HD-texture path together, and keeps
the anti-bleed guard that non-windowed 2D draws rely on at upscaled
internal resolutions. vUV is promoted alongside: fp16 loses the sub-texel
fraction above u = 128 and can mis-select texels under upscaling.

Prebuilt SPIR-V regenerated with the same glslc (byte-identical output on
the unchanged shaders); only the TEXTURED primitive/feedback variants
change.

+++ rhi/shaders_vulkan/vram.h
@@ -1,11 +1,17 @@
 #ifndef VRAM_H
 #define VRAM_H
 
-layout(location = 1) in mediump vec2 vUV;
+/* vUV and vTexLimits must be highp. The rhi sends UINT16_MAX limits to
+ * disable clamping under a texture window; a driver honouring
+ * RelaxedPrecision with 16-bit integers (Adreno, Mali) folds that to -1
+ * and clamp_coord() collapses every texel to base - 1. vUV in fp16 loses
+ * the sub-texel fraction above u = 128, which mis-selects texels under
+ * upscaling. vParam, vBaseUV and vWindow stay within int16 range. */
+layout(location = 1) in highp vec2 vUV;
 layout(location = 2) flat in mediump ivec3 vParam;
 layout(location = 3) flat in mediump ivec2 vBaseUV;
 layout(location = 4) flat in mediump ivec4 vWindow;
-layout(location = 5) flat in mediump ivec4 vTexLimits;
+layout(location = 5) flat in highp ivec4 vTexLimits;
 #if defined(UNSCALED)
 layout(set = 0, binding = 0) uniform mediump usampler2D uFramebuffer;
 #else
